👷(changelog) shift changelog maintenance burden to release manager
Most recent external contributions (including my own!) stumble on
the changelog. IMHO, this represents an unnecessary friction for new
contributors and a loss of focus for the reviewing team.
I propose that we put the release manager in charge of maintaining
the CHANGELOG.md file at each release, based on the commits
being included in that specific release. To view unreleased changes,
everyone can just browse the latest commit history.