⚗️(front) conditionaly install blocknote xl packages
The blocknote xl packages are licensed under AGPL v3. We want to allow a
reuser to build the project without them. For this, in a first step we
don't list these packages in the dependencies but install them using a
postinstall script.
This script is a node script looking in every package.json files under
the apps directory, look if an `extraDependencies` key exists and if yes
hten will install the listed dependencies after prompting the user if
wants or not to install them.
Also an environment variable AUTO_INSTALL_EXTRA_DEPS can be used to
explicitly install all the dependencies or to explicitly not install
them. This will help to build an image with and without these
dependencies. Moreover we also need thisin a CI context.